Modularize, typo fix #22
